requirements.txt
app==0.0.1
config==0.5.1
fastapi==0.115.8
python-dotenv==1.0.1
Requests==2.32.3
uvicorn==0.34.0
